chore: Split specs into unit and integration
