+++ test/integration/support/protractor-config/protractor-sharded.conf.js
