+++ test/integration/support/protractor-config/protractor-with-suite.conf.js
