+++ test/unit/support/fixtures/failed-jasmine2-test-output.txt
