+++ test/unit/support/fixtures/multicapabilities-withspecs.txt
